cpuidle: lpm-levels: Disable dynamic routing of interrupts
LPM driver routes rpm-smd and mpm interrupts dynamically to the core
that is going to wake up next. For some targets for power reasons
it is recommended to affine these interrupts always to specific cores.

Add option to disable dynamic routing of these interrupts inside LPM
driver.

+++ drivers/cpuidle/lpm-levels.c
@@ -496,7 +496,7 @@ static int cluster_configure(struct lpm_cluster *cluster, int idx,
 	}
 	if (level->notify_rpm) {
 		struct cpumask nextcpu, *cpumask;
-                uint32_t us;
+		uint64_t us;
 
 		us = get_cluster_sleep_time(cluster, &nextcpu, from_idle);
 		cpumask = level->disable_dynamic_routing ? NULL : &nextcpu;
